api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pyjwt==2.9.0
pydantic==2.9.0

